Added mapping.conf
This is a simple lofi mapping file. It maps actual lofi's to the
physical locations that are found in layout.conf. The format may change in
the future.

For now, each lofi is identified by its NodeID. Additionally, letters
are appended that represent different switches on the same node. For
example if a single LoFi is servicing two switches at two nearby
locations.